#huber loss function #' @export huberLossPw <- function(x, hgamma = 0.05) stopifnot(is.double(x)) if (abs(x) <= hgamma) 0.5 * (x) ^ 2 else hgamma * abs(x) - ((hgamma ^ 2) / 2)

#' @export huberLoss <- function(x, hgamma = 0.05) sapply(x, huberLossPw, hgamma = hgamma)

dlnormMode(meanlog, sdlog, shift = 0)

Arguments

meanlog
the meanlog argument for dlnorm.

sdlog
the sdlog argument for dlnorm

shift
the shift parameter modulates the shift of the log-normal density function or 0.

Value

the value of the mode function for the given parameters.

Details

huberLoss2 <- function(x1, x2, hgamma = 0.05) huberLoss(x1 - x2, hgamma = hgamma) Mode function of shifted dlnorm.