Calculates the nonlinear fits for the given tibble. Data will be grouped by combinationId, species, fragment, adduct, polarity, calculatedMass, foundMassRange[ppm], and group columns. Calculates non linear regression models by performing an interative grid search within the coordinate bounds provided by lower and upper vectors, starting at start_lower and start_upper. Intermediate models are scored by the AIC value until convergence has been achieved as to the defaults of nls.multstart.

Calculates the nonlinear fits for the given tibble. Data will be grouped by combinationId, species, fragment, adduct, polarity, calculatedMass, foundMassRange[ppm], and group columns. Calculates non linear regression models by performing an interative grid search within the coordinate bounds provided by lower and upper vectors, starting at start_lower and start_upper. Intermediate models are scored by the AIC value until convergence has been achieved as to the defaults of nls.multstart.

fits(tibble, outputPrefix, group, instrumentId, skipGroupOutput = TRUE,
  start_lower, start_upper, lower, upper,
  minSamplesPerCombinationId = 50)

Arguments

tibble
the data tibble containing fragment data.

outputPrefix
the output prefix for the file.

group
the group that this data belongs to.

instrumentId
the instrumentId of this data.

skipGroupOutput
whether to skip writing of grouping information (for debugging).

start_lower
the lower bound to start the parameter grid search.

start_upper
the upper bound to start the parameter grid search.

lower
the lower bound for the parameter grid search.

upper
the upper bound for the parameter grid search.

minSamplesPerCombinationId
the minimum number of data points required per fragment / adduct / ppm combination to be considered for model calculation.