flipr: Flexible inference via permutations in R

The flipr package provides a flexible permutation framework for making inference such as point estimation, confidence intervals or hypothesis testing, on any kind of data, be it univariate, multivariate, or more complex such as network-valued data, topological data, functional data or density-valued data.

See also

Author

Maintainer: Aymeric Stamm aymeric.stamm@cnrs.fr (ORCID)

Authors:

Alessia Pini alessia.pini@unicatt.it

Simone Vantini simone.vantini@polimi.it

Other contributors:

Juliette Chiapello [contributor]

Manon Simonot manon.simonot@univ-nantes.fr [contributor]