Calculates non linear regression models by performing an interative grid search within the coordinate bounds provided by lower and upper vectors, starting at start_lower and start_upper. Intermediate models are scored by the AIC value until convergence has been achieved as to the defaults of nls.multstart.

Calculates non linear regression models by performing an interative grid search within the coordinate bounds provided by lower and upper vectors, starting at start_lower and start_upper. Intermediate models are scored by the AIC value until convergence has been achieved as to the defaults of nls.multstart.

flip(projectDir = getwd(), plotFormat = "png",
  filePattern = "*_fip.tsv", dataPlots = TRUE,
  minPrecursorCollisionEnergy = 0, start_lower, start_upper, lower,
  upper, trainModel = FALSE, minDataPoints = 0, max_iter = 500)

Arguments

projectDir
the path to the project directory containing the '_fip.tsv' files as input.

plotFormat
the plot format, as supported by ggplot2::ggsave.

filePattern
the file pattern for 'fip' files.

dataPlots
whether data plots (diagnostics) should be created.

minPrecursorCollisionEnergy
the minimum precursor collision energy to consider for model training.

start_lower
the lower bound to start the parameter grid search.

start_upper
the upper bound to start the parameter grid search.

lower
the lower bound for the parameter grid search.

upper
the upper bound for the parameter grid search.

trainModel
whether the model should be calculated.

minDataPoints
the minimum number of data points required per fragment / adduct / ppm combination to be considered for model calculation.

max_iter
the number of combinations for grid expansion starting parameters.

Value

The list of fit tables which includes in named element name the input file name for the model data, in named element fits a list with the named elements fits (nonlinear fits), params (parameters), CI (confidence intervals for params), preds (immediate predictions), nls.tibble.unfiltered (unfiltered input data), nls.tibble (data for calculation of fits), and preds_from_data (predictions from equidistantly resampled x-value range).